Separate concerns
Continuation of WIP for issue #63
